Actually connect to PYX for IRC users.
This does very basic connections to PYX when an IRC user connects. It uses their nick as their username and if that name is already in use it boots them from IRC.

After connecting, the long poll process is started. Any events that come in from that (including noop events) get sent to the IRC user in a raw format as a PRIVMSG from Xyzzy.