Add build version information.
Build with https://github.com/ahmetb/govvv to set everything automatically, or

```go build -ldflags "-X main.GitSummary=foo -X main.GitBranch=bar"```

to do it manually.

GitSummary is expected to be the short sha1 (with suffix -dirty if there are uncomitted changes), and GitBranch is expected to be the name of the branch (usually master).