Try to handle more errors.
I've gotten panics somewhat frequently (sometimes multiple times a day, sometimes days between occurances), which seems to be because we're getting an error page instead of a JSON object. Handle those a bit more directly, and also try to recover in the goroutine that was causing the panic. The direct handling won't really stop the panic, but it'll add more logging.